app/requirements.txt
tensorflow==2.15
tf_keras
streamlit
pillow
numpy
